html/semantics/embedded-content/the-video-element/video-loading-attr-lazy.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="http://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<video id="target" loading="lazy"></video>

<script>
  const video = document.getElementById("target");

  test(() => {
    assert_equals(
      video.loading,
      "lazy",
      "Video loading attribute should be 'lazy' when set to 'lazy'"
    );
  }, "Video loading attribute should reflect the loading content attribute");
</script>
